+++ src/main/java/tn/esprit/spring/AhmedGuedri/Repositories/FeesRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.spring.AhmedGuedri.Repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.spring.AhmedGuedri.entities.Fees;
 
-public interface FeesRepository extends JpaRepository<Fees, Integer> {
+public interface FeesRepository extends JpaRepository<Fees, Long> {
 }
