+++ src/main/java/tn/esprit/spring/AhmedGuedri/Repositories/TaxRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.spring.AhmedGuedri.Repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.spring.AhmedGuedri.entities.Tax;
 
-public interface TaxRepository extends JpaRepository<Tax, Integer> {
+public interface TaxRepository extends JpaRepository<Tax, Long> {
 }
