+++ src/main/java/tn/esprit/spring/AhmedGuedri/Services/IProductService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface IProductService {
     List<Products> retrieveAllProducts();
     Products addOrUpdateProduct(Products p);
-    Products getProduct(Integer IdProducts);
-    void removeProduct(Integer IdProducts);
+    Products getProduct(Long IdProducts);
+    void removeProduct(Long IdProducts);
 }
